refactor: create <PricingPackagesBlueSection />
